requirements.txt
fastapi>=0.103.0,<0.104.0
uvicorn>=0.23.0,<0.24.0
sqlalchemy>=2.0.0,<2.1.0
psycopg2-binary
python-multipart==0.0.6
azure-storage-blob==12.19.0